requirements.txt
flask
python-docx
pdfminer.six
